Move NuGetExtensibilityEtw to NuGet.VisualStudio.Common
So that it's accessible from NuGet.SolutionRestoreManager
