+++ src/NuGet.Clients/NuGet.VisualStudio.Implementation/Extensibility/VsGlobalPackagesInitScriptExecutor.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using NuGet.PackageManagement.VisualStudio;
 using NuGet.Packaging.Core;
 using NuGet.Versioning;
+using NuGet.VisualStudio.Etw;
 using NuGet.VisualStudio.Telemetry;
 
 namespace NuGet.VisualStudio.Implementation.Extensibility
