+++ src/NuGet.Clients/NuGet.VisualStudio.Implementation/Extensibility/VsPackageInstallerProjectEvents.cs
@@ -4,6 +4,7 @@
 using System;
 using System.ComponentModel.Composition;
 using NuGet.PackageManagement;
+using NuGet.VisualStudio.Etw;
 
 namespace NuGet.VisualStudio.Implementation.Extensibility
 {
