+++ src/NuGet.Clients/NuGet.VisualStudio.Implementation/Extensibility/VsPackageUninstaller.cs
@@ -15,6 +15,7 @@ using NuGet.Packaging.PackageExtraction;
 using NuGet.Packaging.Signing;
 using NuGet.ProjectManagement;
 using NuGet.Protocol.Core.Types;
+using NuGet.VisualStudio.Etw;
 using NuGet.VisualStudio.Telemetry;
 
 namespace NuGet.VisualStudio.Implementation.Extensibility
