+++ src/NuGet.Clients/NuGet.VisualStudio.Implementation/Extensibility/VsPathContext.cs
@@ -6,6 +6,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using NuGet.Configuration;
+using NuGet.VisualStudio.Etw;
 using NuGet.VisualStudio.Telemetry;
 
 namespace NuGet.VisualStudio.Implementation.Extensibility
