+++ src/NuGet.Clients/NuGet.VisualStudio.Implementation/Extensibility/VsProjectJsonToPackageReferenceMigrator.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using Microsoft;
 using NuGet.PackageManagement.VisualStudio;
 using NuGet.ProjectManagement.Projects;
+using NuGet.VisualStudio.Etw;
 using NuGet.VisualStudio.Implementation.Resources;
 
 namespace NuGet.VisualStudio.Implementation.Extensibility
